# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Reputation Management System API
  version: 1.0.0
  description: |
    API для системы управления репутацией компании. 
    Система позволяет автоматизировать сбор, анализ и мониторинг отзывов 
    о компаниях из различных источников (Яндекс, Google, 2GIS).
  contact:
    name: Reputation Management System
    email: support@reputation-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
  - url: https://api.reputation-system.com/api
    description: Production сервер
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен аутентификации. Получите токен через /auth/login, затем используйте его в формате: Bearer {token}"
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: password123
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
    JwtResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzUxMiJ9..."
        refreshToken:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        tokenType:
          type: string
          example: "Bearer"
        expiresIn:
          type: integer
          example: 900
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [ADMIN, MANAGER, MODERATOR, SPECIALIST]
        companyId:
          type: integer
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation successful"
...
    CreateCompanyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        industry:
          type: string
          maxLength: 100
        website:
          type: string
          maxLength: 255
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        industry:
          type: string
          maxLength: 100
        website:
          type: string
          maxLength: 255
        isActive:
          type: boolean
    CompanyResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        industry:
          type: string
        website:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SourceResponse:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        platform:
          type: string
          enum: [YANDEX, GOOGLE, TWOGIS]
        name:
          type: string
        url:
          type: string
        isActive:
          type: boolean
        lastCollection:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
        sourceId:
          type: integer
        externalId:
          type: string
        author:
          type: string
        rating:
          type: number
          format: double
          minimum: 1.0
          maximum: 5.0
        text:
          type: string
        publishedAt:
          type: string
          format: date-time
        collectedAt:
          type: string
          format: date-time
        sentiment:
          type: string
          enum: [POSITIVE, NEUTRAL, NEGATIVE]
        createdAt:
          type: string
          format: date-time
...
    CreateCommentRequest:
      type: object
      required:
        - taskId
        - text
      properties:
        taskId:
          type: integer
        text:
          type: string
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
    GenerateReportRequest:
      type: object
      required:
        - companyId
        - type
        - periodStart
        - periodEnd
        - format
      properties:
        companyId:
          type: integer
        type:
          type: string
          enum: [SUMMARY, DETAILED, ANALYTICS]
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        format:
          type: string
          enum: [PDF, EXCEL, CSV]
    ReportResponse:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        createdById:
          type: integer
        createdByName:
          type: string
        type:
          type: string
          enum: [SUMMARY, DETAILED, ANALYTICS]
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        format:
          type: string
          enum: [PDF, EXCEL, CSV]
        filePath:
          type: string
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - Аутентификация
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токенов (access и refresh)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '401':
          description: Неверные учетные данные
        '400':
          description: Ошибка валидации
  /auth/refresh:
    post:
      tags:
        - Аутентификация
      summary: Обновление токена
      description: Обновление access токена с помощью refresh токена
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Токен успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '403':
          description: Refresh токен истек или не найден
        '400':
          description: Ошибка валидации
  /auth/logout:
    post:
      tags:
        - Аутентификация
      summary: Выход из системы
      description: Выход из системы и удаление refresh токена
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Успешный выход из системы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Ошибка валидации
...
  /companies:
    post:
      tags:
        - Управление компаниями
      summary: Создать компанию
      description: Создание новой компании в системе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Компания успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          description: Ошибка валидации
        '403':
          description: Недостаточно прав доступа
    get:
      tags:
        - Управление компаниями
      summary: Получить все компании
      description: Получение списка всех компаний
      responses:
        '200':
          description: Список компаний
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyResponse'
  /companies/{id}:
    get:
      tags:
        - Управление компаниями
      summary: Получить компанию по ID
      description: Получение информации о компании по идентификатору
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о компании
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '404':
          description: Компания не найдена
    put:
      tags:
        - Управление компаниями
      summary: Обновить компанию
      description: Обновление информации о компании
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: Компания успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '404':
          description: Компания не найдена
    delete:
      tags:
        - Управление компаниями
      summary: Удалить компанию
      description: Удаление компании из системы
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Компания успешно удалена
        '404':
          description: Компания не найдена
  /companies/{id}/statistics:
    get:
      tags:
        - Управление компаниями
      summary: Получить статистику компании
      description: Получение статистики по компании
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Статистика компании
          content:
            application/json:
              schema:
                type: object
...
  /sources/company/{companyId}:
    get:
      tags:
        - Управление источниками
      summary: Получить источники компании
      description: Получение списка источников для конкретной компании
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список источников компании
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceResponse'
...
  /reviews/company/{companyId}:
    get:
      tags:
        - Работа с отзывами
      summary: Получить отзывы компании
      description: Получение списка отзывов для конкретной компании
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список отзывов компании
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
...
  /comments:
    post:
      tags:
        - Работа с комментариями
      summary: Создать комментарий
      description: Создание комментария к задаче
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Комментарий успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
  /comments/task/{taskId}:
    get:
      tags:
        - Работа с комментариями
      summary: Получить комментарии к задаче
      description: Получение списка комментариев для конкретной задачи
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
  /comments/{id}:
    delete:
      tags:
        - Работа с комментариями
      summary: Удалить комментарий
      description: Удаление комментария из системы
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Комментарий успешно удален
        '404':
          description: Комментарий не найден
  /reports/generate:
    post:
      tags:
        - Управление отчетами
      summary: Сгенерировать отчет
      description: Генерация отчета по указанным параметрам
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateReportRequest'
      responses:
        '201':
          description: Отчет успешно сгенерирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
...
  /reports/{id}:
    get:
      tags:
        - Управление отчетами
      summary: Получить отчет по ID
      description: Получение информации об отчете по идентификатору
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об отчете
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Отчет не найден
    delete:
      tags:
        - Управление отчетами
      summary: Удалить отчет
      description: Удаление отчета из системы
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Отчет успешно удален
        '404':
          description: Отчет не найден
  /reports/{id}/download:
    get:
      tags:
        - Управление отчетами
      summary: Скачать отчет
      description: Скачивание сгенерированного отчета
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Файл отчета
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Отчет не найден
...